Cleanup file lengths
